refactor: 탭 상태 관리 훅으로 개선
